pyproject.toml
[tool.poetry]
name = "chatgpt_dnac_analyzer"
version = "0.1.2"
description = "A chatGPT powered Cisco DNAC analyzer"
authors = ["John Capobianco <ptcapo@gmail.com>"]
packages = [{ include = "chatgpt_dnac_analyzer", from = "src"}]
readme = "README.md"

[tool.poetry.scripts]
chatgpt_dnac_analyzer = "chatgpt_dnac_analyzer.script:run"

[tool.poetry.dependencies]
python = "^3.8"
rich-click = "^1.6.1"
openai = "^0.27.2"
python-dotenv = "^1.0.0"
requests = "^2.28.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
